requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-storage
Pillow